requirements.txt
backcall==0.1.0
bleach==2.1.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
decorator==4.3.0
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.41.0
django-countries==6.1.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
entrypoints==0.2.3
html5lib==1.0.1
idna==2.9
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.5.0
numpy==1.14.5
oauthlib==3.1.0
pandas==0.23.1
pandocfilters==1.4.2
parso==0.2.1
pickleshare==0.7.4
Pillow==7.0.0
prompt-toolkit==1.0.15
psycopg2==2.8.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
pywinpty==0.5.4
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.23.0
requests-oauthlib==1.3.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.1
sympy==1.1.1
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.25.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.53.0
widgetsnbextension==3.2.1
